requirements.txt
requests
pandas
beautifulsoup4
google-generativeai
google-api-python-client
google-auth-oauthlib
pytz 